Extract relation reference rendering
